requirements.txt
apache-airflow
apache-airflow-providers-docker
psycopg2-binary
pydantic-settings==2.2.1